# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Training Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trainings:
    get:
      operationId: listTrainings
      tags: [trainings]
      summary: List trainings visible to the current user
      security:
        - cookieAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated list of trainings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedTrainings"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      operationId: createTraining
      tags: [trainings]
      summary: Create a new training (training_lead+)
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingRequest"
      responses:
        "201":
          description: Training created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}:
    get:
      operationId: getTraining
      tags: [trainings]
      summary: Get a training with its content items
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      responses:
        "200":
          description: Training detail with content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDetailResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateTraining
      tags: [trainings]
      summary: Update a training (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrainingRequest"
      responses:
        "200":
          description: Updated training
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteTraining
      tags: [trainings]
      summary: Soft-delete a training (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      responses:
        "200":
          description: Training deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/content:
    post:
      operationId: addTrainingContent
      tags: [trainings]
      summary: Add a YouTube or Slides URL content item (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrainingContentRequest"
      responses:
        "201":
          description: Content item added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingContentResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/content/scorm:
    post:
      operationId: uploadScorm
      tags: [trainings]
      summary: Upload a SCORM package ZIP file (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: SCORM content uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingContentResponse"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/content/pptx:
    post:
      operationId: uploadPptx
      tags: [trainings]
      summary: Upload a PowerPoint (.pptx) file (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: PPTX content uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingContentResponse"
        "400":
          description: Invalid file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/content/{contentId}:
    delete:
      operationId: deleteTrainingContent
      tags: [trainings]
      summary: Delete a training content item (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
        - in: path
          name: contentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Content deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trainings/{id}/scorm-complete:
    post:
      operationId: scormComplete
      tags: [trainings]
      summary: Record SCORM completion event
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                score:
                  type: integer
      responses:
        "200":
          description: Completion recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /trainings/{id}/assignments:
    get:
      operationId: listTrainingAssignments
      tags: [trainings]
      summary: List group assignments for a training (training_lead+)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      responses:
        "200":
          description: List of group assignments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingAssignmentListResponse"

    post:
      operationId: assignTrainingToGroup
      tags: [trainings]
      summary: Assign a training to a group (admin only)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignToGroupRequest"
      responses:
        "201":
          description: Assignment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingAssignmentResponse"
        "409":
          description: Already assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trainings/{id}/assignments/{assignmentId}:
    delete:
      operationId: unassignTrainingFromGroup
      tags: [trainings]
      summary: Remove a training group assignment (admin only)
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/TrainingId"
        - in: path
          name: assignmentId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assignment removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ─── Quizzes ─────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
...
    TrainingId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Training UUID
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    TrainingContentType:
      type: string
      enum: [scorm, youtube, slides, pptx]
    Training:
      type: object
      required: [id, title, isActive, createdAt, updatedAt]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
          nullable: true
        estimatedDurationMinutes:
          type: integer
          nullable: true
        isActive:
          type: boolean
        createdBy:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TrainingContent:
      type: object
      required: [id, trainingId, type, createdAt]
      properties:
        id:
          type: string
        trainingId:
          type: string
        type:
          $ref: "#/components/schemas/TrainingContentType"
        title:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        filePath:
          type: string
          nullable: true
        displayOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
    TrainingGroupAssignment:
      type: object
      required: [id, trainingId, groupId, assignedAt]
      properties:
        id:
          type: string
        trainingId:
          type: string
        groupId:
          type: string
        dueDate:
          type: string
          format: date-time
          nullable: true
        assignedBy:
          type: string
          nullable: true
        assignedAt:
          type: string
          format: date-time
...
    CreateTrainingRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
        description:
          type: string
        estimatedDurationMinutes:
          type: integer
    UpdateTrainingRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        estimatedDurationMinutes:
          type: integer
        isActive:
          type: boolean
    AddTrainingContentRequest:
      type: object
      required: [type, url]
      properties:
        type:
          type: string
          enum: [youtube, slides]
        title:
          type: string
        url:
          type: string
        displayOrder:
          type: integer
    AssignToGroupRequest:
      type: object
      required: [groupId]
      properties:
        groupId:
          type: string
        dueDate:
          type: string
          format: date-time
...
    TrainingResponse:
      type: object
      required: [training]
      properties:
        training:
          $ref: "#/components/schemas/Training"
    TrainingDetailResponse:
      type: object
      required: [training, content]
      properties:
        training:
          $ref: "#/components/schemas/Training"
        content:
          type: array
          items:
            $ref: "#/components/schemas/TrainingContent"
    TrainingContentResponse:
      type: object
      required: [content]
      properties:
        content:
          $ref: "#/components/schemas/TrainingContent"
    PaginatedTrainings:
      type: object
      required: [trainings, total, page, limit]
      properties:
        trainings:
          type: array
          items:
            $ref: "#/components/schemas/Training"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    TrainingAssignmentResponse:
      type: object
      required: [assignment]
      properties:
        assignment:
          $ref: "#/components/schemas/TrainingGroupAssignment"
    TrainingAssignmentListResponse:
      type: object
      required: [assignments]
      properties:
        assignments:
          type: array
          items:
            $ref: "#/components/schemas/TrainingGroupAssignment"
...