# lib/api-zod/src/generated/types/paginatedTrainings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Training } from './training';

export interface PaginatedTrainings {
  trainings: Training[];
  total: number;
  page: number;
  limit: number;
}