# lib/api-zod/src/generated/types/trainingAssignmentResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { TrainingGroupAssignment } from './trainingGroupAssignment';

export interface TrainingAssignmentResponse {
  assignment: TrainingGroupAssignment;
}