# lib/api-zod/src/generated/types/trainingDetailResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Training Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Training } from './training';
import type { TrainingContent } from './trainingContent';

export interface TrainingDetailResponse {
  training: Training;
  content: TrainingContent[];
}